Bolster help and add Cargo flags
This commit explicitly takes many flags that Cargo takes to filter the package
selection and otherwise configure what packages are being built or what targets
are being built. The intention here is to take arguments but forward them all to
the underlying `cargo check`, but doing so in a first-class fashion to avoid
`--` if possible.