accept multi-part solutions/substitutions
This is meant to be the resolution to #141 (and, in a way,
rust-lang/rust#53934). We add two test fixtures demonstrating usage
and correctness.

The JSON comes from a locally-built non-master compiler featuring the
change to the JSON emitter to support this. (Actually, one is from
`--error-format pretty-json` and the other is from `--error-format
json` and subsequent auto-formatting, but the difference in
inter-string-literal newlines doesn't seem to matter.)